Added functional tests and upgrade Hamcrest
